tourn test

+++ Content_prod/logics/modules/gh_push.py
@@ -61,10 +61,10 @@ def gh_push(main_mod, file_dir, file_name, file_content):
                 if f.read() != file_content:    # если содержание меняется
                     repo.create_file(path+file_name, "add "+file_name, file_content, branch="master")
         # для остальных каталогов
-        elif file_content == 'delete':      # если требуется удалить файл
+        elif type(file_content) is str and file_content == 'delete':      # если требуется удалить файл
             contents = repo.get_contents(path+file_name, ref="master")
             repo.delete_file(contents.path, "remove "+file_name, contents.sha, branch="master")
-        elif file_content[:6] == 'rename':      # если требуется переименовать файл
+        elif type(file_content) is str and file_content[:6] == 'rename':      # если требуется переименовать файл
             new_name = file_content[7:] # новое имя файла
             # извлечь его содержимое, удалить текущий и создать с новым именем
             if file_name[-3:] == 'txt':
